organize into snapshotsupport package
