+++ service/src/main/java/org/databiosphere/workspacedataservice/dataimport/snapshotsupport/WsmSnapshotSupport.java
@@ -1,4 +1,4 @@
-package org.databiosphere.workspacedataservice.dataimport;
+package org.databiosphere.workspacedataservice.dataimport.snapshotsupport;
 
 import bio.terra.datarepo.model.RelationshipModel;
 import bio.terra.datarepo.model.SnapshotModel;
