+++ service/src/test/java/org/databiosphere/workspacedataservice/dataimport/snapshotsupport/WdsSnapshotSupportTest.java
@@ -1,4 +1,4 @@
-package org.databiosphere.workspacedataservice.dataimport;
+package org.databiosphere.workspacedataservice.dataimport.snapshotsupport;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
