+++ service/src/test/java/org/databiosphere/workspacedataservice/pact/WsmPactTest.java
@@ -25,7 +25,7 @@ import com.google.common.collect.ImmutableMap;
 import java.util.Arrays;
 import java.util.UUID;
 import org.databiosphere.workspacedataservice.activitylog.ActivityLogger;
-import org.databiosphere.workspacedataservice.dataimport.WsmSnapshotSupport;
+import org.databiosphere.workspacedataservice.dataimport.snapshotsupport.WsmSnapshotSupport;
 import org.databiosphere.workspacedataservice.retry.RestClientRetry;
 import org.databiosphere.workspacedataservice.sam.BearerTokenFilter;
 import org.databiosphere.workspacedataservice.shared.model.WorkspaceId;
